Create seed data and Update Serializers to reflect needed data

+++ db/migrate/20230925220634_create_ingredients.rb
@@ -3,7 +3,7 @@ class CreateIngredients < ActiveRecord::Migration[7.0]
     create_table :ingredients do |t|
       t.string :name
       t.string :unit
-      t.integer :qty
+      t.decimal :qty
       t.references :recipe, null: false, foreign_key: true
 
       t.timestamps
